# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Milo API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/ui/bootstrap": {
      "get": {
        "tags": [
          "UI bootstrap",
          "client configuration"
        ],
        "summary": "Bootstrap",
        "operationId": "bootstrap_ui_bootstrap_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "name": "conversation_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Conversation Limit"
            }
          },
          {
            "name": "conversation_cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 256
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui/resolve": {
      "get": {
        "tags": [
          "UI bootstrap",
          "client configuration"
        ],
        "summary": "Resolve Ui Reference",
        "description": "Resolve an authenticated deep link without accepting send authority.",
        "operationId": "resolve_ui_reference_ui_resolve_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "conversation",
                "action",
                "draft",
                "memory",
                "connection",
                "job",
                "task"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schedules/resolve-time": {
      "post": {
        "tags": [
          "UI bootstrap",
          "client configuration"
        ],
        "summary": "Resolve Schedule Time",
        "operationId": "resolve_schedule_time_schedules_resolve_time_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveTimeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/owner-drafts": {
      "post": {
        "tags": [
          "UI bootstrap",
          "client configuration"
        ],
        "summary": "Owner Draft",
        "operationId": "owner_draft_conversations__conversation_id__owner_drafts_post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 8,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerDraftInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ui/bootstrap": {
      "get": {
        "tags": [
          "UI bootstrap",
          "client configuration"
        ],
        "summary": "Bootstrap",
        "operationId": "bootstrap_v1_ui_bootstrap_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "name": "conversation_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Conversation Limit"
            }
          },
          {
            "name": "conversation_cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 256
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ui/resolve": {
      "get": {
        "tags": [
          "UI bootstrap",
          "client configuration"
        ],
        "summary": "Resolve Ui Reference",
        "description": "Resolve an authenticated deep link without accepting send authority.",
        "operationId": "resolve_ui_reference_v1_ui_resolve_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "conversation",
                "action",
                "draft",
                "memory",
                "connection",
                "job",
                "task"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schedules/resolve-time": {
      "post": {
        "tags": [
          "UI bootstrap",
          "client configuration"
        ],
        "summary": "Resolve Schedule Time",
        "operationId": "resolve_schedule_time_v1_schedules_resolve_time_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveTimeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OwnerDraftInput": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Text"
          },
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "expected_control_epoch": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Control Epoch"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "OwnerDraftInput"
      },
...
      "ResolveTimeInput": {
        "properties": {
          "local_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Local Datetime"
          },
          "timezone": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Timezone"
          },
          "ambiguity_policy": {
            "type": "string",
            "enum": [
              "reject",
              "earlier",
              "later"
            ],
            "title": "Ambiguity Policy",
            "default": "reject"
          },
          "gap_policy": {
            "type": "string",
            "const": "reject",
            "title": "Gap Policy",
            "default": "reject"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "local_datetime",
          "timezone"
        ],
        "title": "ResolveTimeInput"
      },
...
    }
  }
}